Updates starter script and error page icons
Updates the install command in the starter page.

Changes the default icon for 'home' to 'House' and updates icons
on the error page for better consistency and clarity.

Adds functionality to remove the default +page.svelte file and create
a new one in the (portal) routes when using the starter script.