Enhances site configuration and UI elements
Allows for optional `alt` attribute for logo images in the footer and header, falling back to the site title if not provided, and adds an `appendTitle` option to the header logo.

Removes " ₡ore" suffix from title in header and footer.

Fixes minor UI issues with icon spacing.

Disables authentication by default.